+++ packages/constants/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @phantom/constants
 
+## 2.0.1
+
+### Patch Changes
+
+- f59f462: Don't fetch all wallets
+- Updated dependencies [f59f462]
+  - @phantom/sdk-types@2.0.1
+
 ## 2.0.0-beta.0
 
 ### Major Changes